Moving mas Raden recursive func code to function page, and set code to dataype set page in playground and removing content.swift
