+++ BinarAcademy.playground/Pages/Function.xcplaygroundpage/Contents.swift
@@ -30,3 +30,39 @@ func convert(string: String) -> Int {
 
 func convert(int: Int) {
 }
+
+
+// Recursive function -> fungsi yang memanggil dirinya sendiri
+// Example find factorial
+// Using Looping
+
+func loopFactorial() -> Int {
+    var number: Int = 7
+    var result: Int = 1
+    
+    if number <= 0 {
+        return 1
+    }
+    
+    while number >= 1 {
+        result *= number
+        number = number - 1
+    }
+    
+    return result
+    
+}
+
+print(loopFactorial())
+
+// Using Recursive
+var number: Int = 7
+
+func recursiveFactorial(_ number: Int) -> Int {
+    if number <= 0 {
+        return 1
+    } else {
+        return number * recursiveFactorial(number - 1)
+    }
+}
+print(recursiveFactorial(number))
