+++ BinarAcademy.playground/contents.xcplayground
@@ -18,5 +18,8 @@
         <page name='Optional'/>
         <page name='Subscript'/>
         <page name='SwitchCase'/>
+        <page name='Inheritance'/>
+        <page name='Protocol'/>
+        <page name='DataType Set'/>
     </pages>
 </playground>
